+++ CRM/Nihrbackbone/Upgrader.php
@@ -497,6 +497,52 @@ class CRM_Nihrbackbone_Upgrader extends CRM_Nihrbackbone_Upgrader_Base {
   }
 
   /**
+   * Upgrade 1170 - add entity NbrStudyResearcher table (see https://www.wrike.com/open.htm?id=933254901)
+   *              - migrate existing study researchers
+   *              - remove existing custom field
+   *
+   * @return bool
+   * @throws
+   */
+  public function upgrade_1170(): bool {
+    $this->ctx->log->info(E::ts('Applying update 1170 - add table for entity NbrStudyResearcher'));
+    // create new table for entity
+    if (!CRM_Core_DAO::checkTableExists('civicrm_nbr_study_researcher')) {
+      $create = "CREATE TABLE `civicrm_nbr_study_researcher` (`id` int unsigned NOT NULL AUTO_INCREMENT COMMENT 'Unique NbrStudyResearcher ID',
+         `researcher_contact_id` int unsigned COMMENT 'FK to Contact', `nbr_study_id` int unsigned COMMENT 'FK to Campaign', PRIMARY KEY (`id`),
+         CONSTRAINT FK_civicrm_nbr_study_researcher_researcher_contact_id FOREIGN KEY (`researcher_contact_id`) REFERENCES `civicrm_contact`(`id`) ON DELETE CASCADE,
+         CONSTRAINT FK_civicrm_nbr_study_researcher_nbr_study_id FOREIGN KEY (`nbr_study_id`) REFERENCES `civicrm_campaign`(`id`) ON DELETE CASCADE)
+    ENGINE=InnoDB;";
+      CRM_Core_DAO::executeQuery($create);
+      // migrate existing researchers
+      $dao = CRM_Core_DAO::executeQuery("SELECT nsd_researcher, entity_id AS study_id FROM civicrm_value_nbr_study_data WHERE nsd_researcher IS NOT NULL");
+      while ($dao->fetch()) {
+        if ($dao->study_id && $dao->nsd_researcher) {
+          CRM_Nihrbackbone_BAO_NbrStudyResearcher::createStudyResearcher((int) $dao->study_id, (int) $dao->nsd_researcher);
+        }
+      }
+      // disable existing custom field (to be removed once all is done on live)
+      try {
+        $customFieldId = \Civi\Api4\CustomField::get()
+          ->addSelect('id')
+          ->addWhere('custom_group_id:name', '=', 'nbr_study_data')
+          ->addWhere('name', '=', 'nsd_researcher')
+          ->setLimit(1)
+          ->execute()->first();
+        if ($customFieldId) {
+          \Civi\Api4\CustomField::update()
+            ->addWhere('id', '=', $customFieldId)
+            ->addValue('is_active', FALSE)
+            ->execute();
+        }
+      }
+      catch (API_Exception $ex) {
+      }
+    }
+    return TRUE;
+  }
+
+  /**
    * Swap values for unable to willing columns
    *
    * @param $columns