+++ nihrbackbone.php
@@ -275,15 +275,6 @@ function nihrbackbone_civicrm_config(&$config) {
 }
 
 /**
- * Implements hook_civicrm_xmlMenu().
- *
- * @link http://wiki.civicrm.org/confluence/display/CRMDOC/hook_civicrm_xmlMenu
- */
-function nihrbackbone_civicrm_xmlMenu(&$files) {
-  _nihrbackbone_civix_civicrm_xmlMenu($files);
-}
-
-/**
  * Implements hook_civicrm_install().
  *
  * @link http://wiki.civicrm.org/confluence/display/CRMDOC/hook_civicrm_install
@@ -293,96 +284,12 @@ function nihrbackbone_civicrm_install() {
 }
 
 /**
- * Implements hook_civicrm_postInstall().
- *
- * @link http://wiki.civicrm.org/confluence/display/CRMDOC/hook_civicrm_postInstall
- */
-function nihrbackbone_civicrm_postInstall() {
-  _nihrbackbone_civix_civicrm_postInstall();
-}
-
-/**
- * Implements hook_civicrm_uninstall().
- *
- * @link http://wiki.civicrm.org/confluence/display/CRMDOC/hook_civicrm_uninstall
- */
-function nihrbackbone_civicrm_uninstall() {
-  _nihrbackbone_civix_civicrm_uninstall();
-}
-
-/**
  * Implements hook_civicrm_enable().
  *
  * @link http://wiki.civicrm.org/confluence/display/CRMDOC/hook_civicrm_enable
  */
 function nihrbackbone_civicrm_enable() {
   _nihrbackbone_civix_civicrm_enable();
-}
-
-/**
- * Implements hook_civicrm_disable().
- *
- * @link http://wiki.civicrm.org/confluence/display/CRMDOC/hook_civicrm_disable
- */
-function nihrbackbone_civicrm_disable() {
-  _nihrbackbone_civix_civicrm_disable();
-}
-
-/**
- * Implements hook_civicrm_upgrade().
- *
- * @link http://wiki.civicrm.org/confluence/display/CRMDOC/hook_civicrm_upgrade
- */
-function nihrbackbone_civicrm_upgrade($op, CRM_Queue_Queue $queue = NULL) {
-  return _nihrbackbone_civix_civicrm_upgrade($op, $queue);
-}
-
-/**
- * Implements hook_civicrm_managed().
- *
- * Generate a list of entities to create/deactivate/delete when this module
- * is installed, disabled, uninstalled.
- *
- * @link http://wiki.civicrm.org/confluence/display/CRMDOC/hook_civicrm_managed
- */
-function nihrbackbone_civicrm_managed(&$entities) {
-  _nihrbackbone_civix_civicrm_managed($entities);
-}
-
-/**
- * Implements hook_civicrm_caseTypes().
- *
- * Generate a list of case-types.
- *
- * Note: This hook only runs in CiviCRM 4.4+.
- *
- * @link http://wiki.civicrm.org/confluence/display/CRMDOC/hook_civicrm_caseTypes
- */
-function nihrbackbone_civicrm_caseTypes(&$caseTypes) {
-  _nihrbackbone_civix_civicrm_caseTypes($caseTypes);
-}
-
-/**
- * Implements hook_civicrm_angularModules().
- *
- * Generate a list of Angular modules.
- *
- * Note: This hook only runs in CiviCRM 4.5+. It may
- * use features only available in v4.6+.
- *
- * @link http://wiki.civicrm.org/confluence/display/CRMDOC/hook_civicrm_angularModules
- */
-function nihrbackbone_civicrm_angularModules(&$angularModules) {
-  _nihrbackbone_civix_civicrm_angularModules($angularModules);
-}
-
-/**
- * Implements hook_civicrm_alterSettingsFolders().
- *
- * @link http://wiki.civicrm.org/confluence/display/CRMDOC/hook_civicrm_alterSettingsFolders
- */
-function nihrbackbone_civicrm_alterSettingsFolders(&$metaDataFolders = NULL) {
-  _nihrbackbone_civix_civicrm_alterSettingsFolders($metaDataFolders);
 }
 
 /**
@@ -396,8 +303,19 @@ function nihrbackbone_civicrm_entityTypes(&$entityTypes) {
   _nihrbackbone_civix_civicrm_entityTypes($entityTypes);
 }
 
-// --- Functions below this ship commented out. Uncomment as required. ---
+/**
+ * Function to mitigate against this bug: https://github.com/totten/civix/issues/244
+ *
+ * @param $files
+ * @return void
+ */
+function nihrbackbone_civicrm_xmlMenu(&$files) {
+  foreach (glob(__DIR__ . '/xml/Menu/*.xml') as $file) {
+    $files[] = $file;
+  }
+}
 
+// --- Functions below this ship commented out. Uncomment as required. ---
 
 /**
  * Implements hook_civicrm_navigationMenu().
@@ -415,4 +333,3 @@ function nihrbackbone_civicrm_navigationMenu(&$menu) {
   ));
   _nihrbackbone_civix_navigationMenu($menu);
 } // */
-