+++ sql/auto_install.sql
@@ -30,6 +30,7 @@
 
 SET FOREIGN_KEY_CHECKS=0;
 
+DROP TABLE IF EXISTS `civicrm_nbr_study_researcher`;
 DROP TABLE IF EXISTS `civicrm_nbr_mailing`;
 DROP TABLE IF EXISTS `civicrm_nbr_import_log`;
 DROP TABLE IF EXISTS `civicrm_nbr_county`;
@@ -53,16 +54,16 @@ CREATE TABLE `civicrm_nbr_county` (
 
      `id` int unsigned NOT NULL AUTO_INCREMENT  COMMENT 'Unique NbrCounty ID',
      `state_province_id` int unsigned    COMMENT 'FK to State/Province',
-     `synonym` varchar(256)    COMMENT 'County synonym'
+     `synonym` varchar(256)    COMMENT 'County synonym' 
 ,
         PRIMARY KEY (`id`)
-
-    ,     INDEX `index_synonym`(
+ 
+    ,     UNIQUE INDEX `index_synonym`(
         synonym
   )
-
-,          CONSTRAINT FK_civicrm_nbr_county_state_province_id FOREIGN KEY (`state_province_id`) REFERENCES `civicrm_state_province`(`id`) ON DELETE CASCADE
-)    ;
+  
+,          CONSTRAINT FK_civicrm_nbr_county_state_province_id FOREIGN KEY (`state_province_id`) REFERENCES `civicrm_state_province`(`id`) ON DELETE CASCADE  
+)  ENGINE=InnoDB  ;
 
 -- /*******************************************************
 -- *
@@ -79,13 +80,13 @@ CREATE TABLE `civicrm_nbr_import_log` (
      `filename` varchar(128)    COMMENT 'Name of the import file that is being logged',
      `message_type` varchar(128)    COMMENT 'Type of message (info, warning, error)',
      `message` text    COMMENT 'Message',
-     `logged_date` date    COMMENT 'The date the message was logged'
+     `logged_date` date    COMMENT 'The date the message was logged' 
 ,
         PRIMARY KEY (`id`)
-
-
-
-)    ;
+ 
+ 
+ 
+)  ENGINE=InnoDB  ;
 
 -- /*******************************************************
 -- *
@@ -101,11 +102,32 @@ CREATE TABLE `civicrm_nbr_mailing` (
      `mailing_id` int unsigned    COMMENT 'FK to Mailing',
      `group_id` int unsigned    COMMENT 'FK to Group',
      `study_id` int unsigned    COMMENT 'FK to Study (Campaign)',
-     `nbr_mailing_type` varchar(32)    COMMENT 'Type of Mailing (invite initially)'
+     `nbr_mailing_type` varchar(32)    COMMENT 'Type of Mailing (invite initially)' 
 ,
         PRIMARY KEY (`id`)
+ 
+ 
+,          CONSTRAINT FK_civicrm_nbr_mailing_mailing_id FOREIGN KEY (`mailing_id`) REFERENCES `civicrm_mailing`(`id`) ON DELETE CASCADE,          CONSTRAINT FK_civicrm_nbr_mailing_group_id FOREIGN KEY (`group_id`) REFERENCES `civicrm_group`(`id`) ON DELETE SET NULL,          CONSTRAINT FK_civicrm_nbr_mailing_study_id FOREIGN KEY (`study_id`) REFERENCES `civicrm_campaign`(`id`) ON DELETE CASCADE  
+)  ENGINE=InnoDB  ;
+
+-- /*******************************************************
+-- *
+-- * civicrm_nbr_study_researcher
+-- *
+-- * Table to link researcher(s) to study
+-- *
+-- *******************************************************/
+CREATE TABLE `civicrm_nbr_study_researcher` (
 
 
-,          CONSTRAINT FK_civicrm_nbr_mailing_mailing_id FOREIGN KEY (`mailing_id`) REFERENCES `civicrm_mailing`(`id`) ON DELETE CASCADE,          CONSTRAINT FK_civicrm_nbr_mailing_group_id FOREIGN KEY (`group_id`) REFERENCES `civicrm_group`(`id`) ON DELETE SET NULL,          CONSTRAINT FK_civicrm_nbr_mailing_study_id FOREIGN KEY (`study_id`) REFERENCES `civicrm_campaign`(`id`) ON DELETE CASCADE
-)    ;
+     `id` int unsigned NOT NULL AUTO_INCREMENT  COMMENT 'Unique NbrStudyResearcher ID',
+     `researcher_contact_id` int unsigned    COMMENT 'FK to Contact',
+     `nbr_study_id` int unsigned    COMMENT 'FK to Campaign' 
+,
+        PRIMARY KEY (`id`)
+ 
+ 
+,          CONSTRAINT FK_civicrm_nbr_study_researcher_researcher_contact_id FOREIGN KEY (`researcher_contact_id`) REFERENCES `civicrm_contact`(`id`) ON DELETE CASCADE,          CONSTRAINT FK_civicrm_nbr_study_researcher_nbr_study_id FOREIGN KEY (`nbr_study_id`) REFERENCES `civicrm_campaign`(`id`) ON DELETE CASCADE  
+)  ENGINE=InnoDB  ;
 
+