+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.44
+* allow multiple researcher for study (see https://www.wrike.com/open.htm?id=933254901)
+>* add new entity NbrStudyResearcher
+>* cater for new entity in study form
+>
+
 ## 2.43
 * set study prevent upload to portal field to default TRUE when adding on the study form